+++ test/schemes/fluid/weakly_compressible_sph/density_diffusion.jl
@@ -4,7 +4,15 @@
         @trixi_testset "show" begin
             density_diffusion = DensityDiffusionAntuono(delta=0.1)
 
-            @test repr(density_diffusion) == "DensityDiffusionAntuono(0.1)"
+            @test repr(density_diffusion) ==
+                  "DensityDiffusionAntuono(delta=0.1, update_everystage=true)"
+
+            callback_density_diffusion = DensityDiffusionAntuono(delta=0.1,
+                                                                 update_everystage=false)
+            @test repr(callback_density_diffusion) ==
+                  "DensityDiffusionAntuono(delta=0.1, update_everystage=false)"
+            @test !TrixiParticles.requires_update_callback(density_diffusion)
+            @test TrixiParticles.requires_update_callback(callback_density_diffusion)
         end
     end
 end